+++ src/apyfixedarray.cc
@@ -1377,6 +1377,29 @@ APyFixedArray APyFixedArray::cast(
     return APyFixedArray(_shape, new_bits, new_int_bits, std::move(result_data));
 }
 
+std::tuple<int, int, std::vector<std::size_t>, std::vector<std::uint64_t>>
+APyFixedArray::python_pickle() const
+{
+    // While pickling, we convert the limb vector to a 64-bit vector, so that the
+    // serialized data becomes consistent between 32-bit and 64-bit systems
+    auto&& u64_vec = limb_vector_to_u64_vec(std::begin(_data), std::end(_data));
+    return std::make_tuple(_bits, _int_bits, _shape, u64_vec);
+}
+
+//! Python un-pickling
+void APyFixedArray::python_unpickle(
+    APyFixedArray* apyfixedarray,
+    const std::tuple<int, int, std::vector<std::size_t>, std::vector<std::uint64_t>>&
+        state
+)
+{
+    auto&& [bits, int_bits, shape, u64_vec] = state;
+    APyFixedArray new_fx(shape, bits, int_bits);
+    new_fx._data = limb_vector_from_u64_vec<APyFixedArray::vector_type>(u64_vec);
+    new_fx._data.resize(bits_to_limbs(bits) * fold_shape(shape));
+    new (apyfixedarray) APyFixedArray(new_fx);
+}
+
 /* ********************************************************************************** *
  * *                      Static conversion from other types                          *
  * ********************************************************************************** */