+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `to_fraction()` and `as_integer_ratio()` for `APyFixed` and `APyFloat`.
 - All scalar classes have a `from_number` method which aliases
 `from_float`/`from_complex` to match the naming in Python 3.14.
+- Support for (de-)serialization of APyTypes' data types using the Python
+  [pickle](https://docs.python.org/3/library/pickle.html) module.
 
 ### Fixed